split reusable WebViewActivity from LocalHelpActivity
