+++ res/layout/web_view_activity.xml
